Skip testing fp16/fp64 types if not supported by a device (#461)
The limitation is that wrapper-like types are not properly handled;
the most important one is zip_iterator

+++ test/parallel_api/algorithm/alg.modifying.operations/swap_ranges.pass.cpp
@@ -131,9 +131,9 @@ int
 main()
 {
     test<wrapper<std::uint16_t>>();
-    test<wrapper<float64_t>>();
+    test<wrapper<float32_t>>();
     test<std::int32_t>();
-    test<float32_t>();
+    test<float64_t>();
 
     return done();
 }
